RFC: Introduce Lanczos filter element
Kwin contains a nice way of downsampling texures in a way that provides
a more appealing result.

This patch exposes the same feature to QtQuick uisng the same algorithm
and settings.

Intended initial usage is the window thumbnails, but I want it available
for more users.

Performance wise, even when using PlasmaCore.WindowThumbnail as a source
and updating constantly usage of the the filter still used <1% of my
CPU. There's almost no work at all on the GUI thread.

It's not done as a pure QML ShaderEffect because we need to upload
uniform arrays and we need to do two passes.

RFC because:
- I haven't attached the glue code as I expect some bikeshedding into
the repo/import name/class name

- Do want the class to be named after how it works (Lanczos filter) or
what it does (SmoothResize)

 - I need to add profile shaders